X-API Refactoring
