+++ src/VirtoCommerce.News.ExperienceApi/Extensions/ServiceCollectionExtensions.cs
@@ -1,6 +1,5 @@
 using GraphQL.MicrosoftDI;
 using Microsoft.Extensions.DependencyInjection;
-using VirtoCommerce.News.ExperienceApi.Services;
 using VirtoCommerce.Xapi.Core.Extensions;
 using VirtoCommerce.Xapi.Core.Infrastructure;
 
@@ -16,9 +15,6 @@ public static class ServiceCollectionExtensions
         });
 
         serviceCollection.AddSingleton<ScopedSchemaFactory<XapiAssemblyMarker>>();
-        serviceCollection.AddSingleton<INewsArticleLocalizationService, NewsArticleLocalizationService>();
-        serviceCollection.AddSingleton<INewsArticleSeoService, NewsArticleSeoService>();
-        serviceCollection.AddSingleton<INewsArticleUserGroupService, NewsArticleUserGroupService>();
 
         return serviceCollection;
     }
