--- src/VirtoCommerce.News.ExperienceApi/Services/INewsArticleSeoService.cs
@@ -1,10 +1,0 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using VirtoCommerce.News.Core.Models;
-
-namespace VirtoCommerce.News.ExperienceApi.Services;
-
-public interface INewsArticleSeoService
-{
-    Task FilterSeoInfosAsync(IList<NewsArticle> newsArticles, string languageCode, string storeId);
-}
